backend/fix/fixes number of rides sent to driver dashboard
